Move rm dir test into correct module
