+++ oxen-rust/src/lib/src/api/client/workspaces/files.rs
@@ -1903,92 +1903,6 @@ mod tests {
         .await
     }
 
-    async fn upload_test_file(
-        remote_repo: &RemoteRepository,
-        fi: &PathBuf,
-    ) -> Result<String, OxenError> {
-        // tokio::fs::create_dir_all(&fi).await?;
-        // tokio::fs::write(&fi, b"Hello world!").await?;
-
-        let writing_to_dir = std::env::temp_dir();
-
-        let parent = fi.parent().unwrap();
-        // println!("creating dir: {}", parent.display());
-        crate::util::fs::create_dir_all(parent)?;
-
-        let file_name = fi.file_name().unwrap();
-        let to_write = writing_to_dir.join(file_name);
-        // println!("writing file: {}", fi.display());
-        // crate::util::fs::write(&fi, b"Hello world!")?;
-        println!("writing file: {}", to_write.display());
-        crate::util::fs::write(&to_write, b"Hello world!")?;
-
-        println!("putting file: {}", fi.display());
-        let commit = crate::api::client::file::put_file(
-            &remote_repo,
-            DEFAULT_BRANCH_NAME,
-            &parent.to_string_lossy(),
-            &to_write,
-            // &fi,
-            // weird(),
-            // None::<String>,
-            Some(&file_name.to_string_lossy()),
-            Some(crate::model::commit::NewCommitBody {
-                message: "a message".into(),
-                author: "a person".into(),
-                email: "email".into(),
-            }),
-        )
-        .await?;
-
-        // println!("removing file: {}", fi.display());
-        // let x = tokio::fs::remove_dir(&fi).await?;
-        // Ok(x)
-        Ok(commit.commit.id)
-    }
-
-    // Test that downloading a non-existent file returns OxenError::PathDoesNotExist
-    #[tokio::test]
-    async fn test_rm_directory() -> Result<(), OxenError> {
-        test::run_remote_created_and_readme_remote_repo_test(|remote_repo| async move {
-            // Remove the committed directory "annotations/train"
-            let annotations_d = Path::new("annotations");
-            let train_d = annotations_d.join("train");
-            let test_d = annotations_d.join("test");
-
-            let train_fi = train_d.join("some_file");
-            let _train_commit = upload_test_file(&remote_repo, &train_fi).await?;
-
-            let test_fi = test_d.join("a_different_file");
-            let test_commit = upload_test_file(&remote_repo, &test_fi).await?;
-
-            let c = api::client::file::delete_file(
-                &remote_repo,
-                DEFAULT_BRANCH_NAME,
-                &train_d,
-                Some(NewCommitBody {
-                    message: "delete".into(),
-                    author: "author".into(),
-                    email: "ox@oxen.ai".into(),
-                }),
-            )
-            .await?;
-
-            println!("deleted {} as commit {}", train_d.display(), c.commit.id);
-
-            let contents =
-                api::client::file::get_file(&remote_repo, DEFAULT_BRANCH_NAME, &train_fi).await;
-            assert!(
-                contents.is_err(),
-                "I should be an error but I'm actually: {:?}",
-                contents
-            );
-
-            Ok(remote_repo)
-        })
-        .await
-    }
-
     #[tokio::test]
     async fn test_stage_file_in_multiple_subdirectories() -> Result<(), OxenError> {
         test::run_remote_repo_test_bounding_box_csv_pushed(|_local_repo, remote_repo| async move {
